+++ src/lib/services/validator.ts
@@ -3,9 +3,18 @@ import axios from "axios";
 import { z } from "zod";
 
 import { CURR_THEME } from "env";
-import { BlockVote, zBig, zUtcDate, zValidatorData } from "lib/types";
 import type { Option, StakingShare, Validator, ValidatorAddr } from "lib/types";
+import {
+  BlockVote,
+  zBig,
+  zCoin,
+  zUtcDate,
+  zValidatorAddr,
+  zValidatorData,
+} from "lib/types";
 import { parseWithError, removeSpecialChars, snakeToCamel } from "lib/utils";
+
+import { zBlocksResponse } from "./block";
 
 interface ValidatorResponse {
   operator_address: ValidatorAddr;
@@ -77,6 +86,30 @@ export const resolveValIdentity = async (
   );
 };
 
+const zHistoricalPowersItem = z.object({
+  hour_rounded_timestamp: zUtcDate,
+  timestamp: zUtcDate,
+  voting_power: zBig,
+});
+
+export const zHistoricalPowersResponse = z
+  .object({
+    items: z.array(zHistoricalPowersItem),
+    total: z.number(),
+  })
+  .transform(snakeToCamel);
+export type HistoricalPowersResponse = z.infer<
+  typeof zHistoricalPowersResponse
+>;
+
+export const getHistoricalPowers = async (
+  endpoint: string,
+  validatorAddr: ValidatorAddr
+): Promise<HistoricalPowersResponse> =>
+  axios
+    .get(`${endpoint}/${validatorAddr}/historical-powers`)
+    .then(({ data }) => parseWithError(zHistoricalPowersResponse, data));
+
 const zValidatorsResponse = z
   .object({
     items: z.array(zValidatorData),
@@ -107,6 +140,18 @@ export const getValidators = async (
       },
     })
     .then(({ data }) => parseWithError(zValidatorsResponse, data));
+
+const zStakingProvisionsResponse = z.object({
+  staking_provisions: zBig,
+});
+export type StakingProvisionsResponse = z.infer<
+  typeof zStakingProvisionsResponse
+>;
+
+export const getValidatorStakingProvisions = async (endpoint: string) =>
+  axios
+    .get(`${endpoint}/staking-provisions`)
+    .then(({ data }) => parseWithError(zStakingProvisionsResponse, data));
 
 const zValidatorDataResponse = z
   .object({
@@ -162,3 +207,76 @@ export const getValidatorUptime = async (
       },
     })
     .then(({ data }) => parseWithError(zValidatorUptimeResponse, data));
+
+const zValidatorDelegationRelatedTxsResponseItem = z
+  .object({
+    tx_hash: z.string(),
+    height: z.number().positive(),
+    tokens: zCoin.array(),
+    timestamp: zUtcDate,
+    validator_address: zValidatorAddr,
+  })
+  .transform(snakeToCamel);
+
+const zValidatorDelegationRelatedTxsResponse = z.object({
+  items: z.array(zValidatorDelegationRelatedTxsResponseItem),
+  total: z.number().nonnegative(),
+});
+
+export type ValidatorDelegationRelatedTxsResponse = z.infer<
+  typeof zValidatorDelegationRelatedTxsResponse
+>;
+
+export const getValidatorDelegationRelatedTxs = async (
+  endpoint: string,
+  validatorAddress: ValidatorAddr,
+  limit: number,
+  offset: number
+) =>
+  axios
+    .get(
+      `${endpoint}/${encodeURIComponent(validatorAddress)}/delegation-related-txs`,
+      {
+        params: {
+          limit,
+          offset,
+        },
+      }
+    )
+    .then(({ data }) =>
+      parseWithError(zValidatorDelegationRelatedTxsResponse, data)
+    );
+
+export const getValidatorProposedBlocks = async (
+  endpoint: string,
+  validatorAddress: ValidatorAddr,
+  limit: number,
+  offset: number
+) =>
+  axios
+    .get(
+      `${endpoint}/${encodeURIComponent(validatorAddress)}/proposed-blocks`,
+      {
+        params: {
+          limit,
+          offset,
+        },
+      }
+    )
+    .then(({ data }) => parseWithError(zBlocksResponse, data));
+
+const zValidatorDelegatorsResponse = z.object({
+  total: z.number().nonnegative(),
+});
+
+export type ValidatorDelegatorsResponse = z.infer<
+  typeof zValidatorDelegatorsResponse
+>;
+
+export const getValidatorDelegators = async (
+  endpoint: string,
+  validatorAddress: ValidatorAddr
+) =>
+  axios
+    .get(`${endpoint}/${encodeURIComponent(validatorAddress)}/delegators`)
+    .then(({ data }) => parseWithError(zValidatorDelegatorsResponse, data));