+++ src/lib/services/validatorService.ts
@@ -12,15 +12,23 @@ import {
 } from "lib/app-provider";
 import type { Nullable, Option, Validator, ValidatorAddr } from "lib/types";
 
+import type { BlocksResponse } from "./block";
 import type {
+  StakingProvisionsResponse,
   ValidatorDataResponse,
+  ValidatorDelegationRelatedTxsResponse,
   ValidatorsResponse,
   ValidatorUptimeResponse,
 } from "./validator";
 import {
+  getHistoricalPowers,
   getValidator,
   getValidatorData,
+  getValidatorDelegationRelatedTxs,
+  getValidatorDelegators,
+  getValidatorProposedBlocks,
   getValidators,
+  getValidatorStakingProvisions,
   getValidatorUptime,
   resolveValIdentity,
 } from "./validator";
@@ -73,6 +81,19 @@ export const useValidatorImage = (
   });
 };
 
+export const useValidatorHistoricalPowers = (validatorAddr: ValidatorAddr) => {
+  const endpoint = useBaseApiRoute("validators");
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.VALIDATOR_HISTORICAL_POWERS, endpoint],
+    async () => getHistoricalPowers(endpoint, validatorAddr),
+    {
+      retry: 1,
+      refetchOnWindowFocus: false,
+    }
+  );
+};
+
 export const useValidators = (
   limit: number,
   offset: number,
@@ -104,6 +125,18 @@ export const useValidators = (
   );
 };
 
+export const useValidatorStakingProvisions = () => {
+  const endpoint = useBaseApiRoute("validators");
+
+  return useQuery<StakingProvisionsResponse>(
+    [CELATONE_QUERY_KEYS.VALIDATOR_STAKING_PROVISIONS, endpoint],
+    async () => getValidatorStakingProvisions(endpoint),
+    {
+      retry: 1,
+    }
+  );
+};
+
 export const useValidatorData = (validatorAddress: ValidatorAddr) => {
   const endpoint = useBaseApiRoute("validators");
 
@@ -128,5 +161,54 @@ export const useValidatorUptime = (
     {
       retry: 1,
     }
+  );
+};
+
+export const useValidatorDelegationRelatedTxs = (
+  validatorAddress: ValidatorAddr,
+  limit: number,
+  offset: number
+) => {
+  const endpoint = useBaseApiRoute("validators");
+
+  return useQuery<ValidatorDelegationRelatedTxsResponse>(
+    [
+      CELATONE_QUERY_KEYS.VALIDATOR_DELEGATION_RELATED_TXS,
+      endpoint,
+      validatorAddress,
+    ],
+    async () =>
+      getValidatorDelegationRelatedTxs(
+        endpoint,
+        validatorAddress,
+        limit,
+        offset
+      ),
+    { retry: 1 }
+  );
+};
+
+export const useValidatorProposedBlocks = (
+  validatorAddress: ValidatorAddr,
+  limit: number,
+  offset: number
+) => {
+  const endpoint = useBaseApiRoute("validators");
+
+  return useQuery<BlocksResponse>(
+    [CELATONE_QUERY_KEYS.VALIDATOR_PROPOSED_BLOCKS, endpoint, validatorAddress],
+    async () =>
+      getValidatorProposedBlocks(endpoint, validatorAddress, limit, offset),
+    { retry: 1 }
+  );
+};
+
+export const useValidatorDelegators = (validatorAddress: ValidatorAddr) => {
+  const endpoint = useBaseApiRoute("validators");
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.VALIDATOR_DELEGATORS, endpoint, validatorAddress],
+    async () => getValidatorDelegators(endpoint, validatorAddress),
+    { retry: 1 }
   );
 };